CrewAI_equipe_creation_BD_candidats/requirements.txt
crewai[tools]
python-dotenv
langchain
agentops
